src/requirements.txt
# pip install torch==2.1.0 torchvision==0.16.0 torchaudio==2.1.0 --index-url https://download.pytorch.org/whl/cu121 xformers
taichi==1.7.0
warp-lang
matplotlib
pyyaml
h5py
einops
timm
trimesh
scikit-image
open3d
PyMCubes
plyfile
transformers==4.47.1
cython
rich
munch
# torch==2.5.0+cu118
torch_cluster
diffusers>=0.32.2
accelerate>=1.1.1
omegaconf
huggingface_hub==0.31.1
tqdm
datasets
xformers
# objaverse_downloader
git+https://github.com/asomoza/image_gen_aux.git
moviepy==1.0.3
hydra-core==1.3.2
gradio==4.44.1
gradio_image_prompter==0.1.0